2-computer-vision-classifier/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
matplotlib>=3.8.0
Pillow>=10.2.0
